==> src/09_scrape-users/05_github-users-to-ctry.Rmd <==
---
title: "International Collaboration Networks"
author: "Brandon L. Kramer"
date: "10/21/2019"
output: html_document
---

This file outlines the process by which we reduced the OSS international data into network format. 

```{r loading nodelist data}
rm(list = ls())

# load packages 
for (pkg in c("tidyverse", "data.table", 
              "tidytable", "countrycode", "RPostgreSQL")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get our data
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the users_gh data (table of all github users) 
oss_data <- dbGetQuery(conn, "SELECT login, location, email, company FROM gh.ctrs_raw_0821")
academic_data <- dbGetQuery(conn, "SELECT login, organization, country, academic FROM gh.sna_ctr_academic_0821;")

# disconnect from postgresql database 
dbDisconnect(conn)
```

```{r}
oss_data %>% 
  drop_na(country_code)
```

## Recoding Country Codes and Writing New Table to SQL ##########################################################################################

This code simply repeats the final recoding step in the country-cleaning.Rmd file. Conceptually, this chunk recodes the location column into a detected_country column and then adds that new information to the already existing country_code column. (Note: Brandon did make a slight adjustment to the US string since it was case sensitive before, which drove our totals down.)

```{r recoding all of the country codes}
# first we have to recode all the strings that could arise in the location colunmn as country codes 
af	<- "\\b(?i)(Afghanistan)\\b"
ax	<- "\\b(?i)(Åland Islands|Aland Islands)\\b"
al	<- "\\b(?i)(Albania)\\b"
dz	<- "\\b(?i)(Algeria)\\b"
as	<- "\\b(?i)(American Samoa)\\b"
ad	<- "\\b(?i)(Andorra)\\b"
ao	<- "\\b(?i)(Angola)\\b"
ai	<- "\\b(?i)(Anguilla)\\b"
aq	<- "\\b(?i)(Antarctica|Antartica)\\b"
ag	<- "\\b(?i)(Antigua|Barbuda)\\b"
ar	<- "\\b(?i)(Argentina|Buenos Aires|La plata)\\b"
am	<- "\\b(?i)(Armenia)\\b"
aw	<- "\\b(?i)(Aruba)\\b"
au	<- "\\b(?i)(Australia|Melbourne(?![, FL|, UK| UK])|Sydney(?!, Nova)|Wollongong|Brisbane|Canberra)\\b"
at  <- "\\b(?i)(Austria|Vienna(?!, V))\\b"
az	<- "\\b(?i)(Azerbaijan)\\b"
bh	<- "\\b(?i)(Bahrain)\\b"
bs	<- "\\b(?i)(Bahamas)\\b"
bd	<- "\\b(?i)(Bangladesh|Dhaka)\\b"
bb	<- "\\b(?i)(Barbados)\\b"
by	<- "\\b(?i)(Belarus)\\b"
be  <- "\\b(?i)(Belgium|Brussels)\\b"
bz	<- "\\b(?i)(Belize)\\b"
bj	<- "\\b(?i)(Benin(?! City))\\b"
bm	<- "\\b(?i)(Bermuda)\\b"
bt	<- "\\b(?i)(Bhutan)\\b"
bo	<- "\\b(?i)(Bolivia)\\b"
bq	<- "\\b(?i)(Bonaire)\\b"
ba	<- "\\b(?i)(Bosnia|Herzegovina)\\b"
bw	<- "\\b(?i)(Botswana)\\b"
bv	<- "\\b(?i)(Bouvet Island)\\b"
br  <- "\\b(?i)(Brazil|Brasil|Aracaju|Sao Paulo|San Paulo|Rio de Janeiro|Brasilia|Brasília|Fortaleza|Araraquara|Barueri|Belo Horizonte|BR,|Brazilian|Campina Grande|Florianpolis|Fortaleza|Goinia-GO|Goinia|Joo Pessoa|Porto Alegre|Rio de |Rio Grande|So Jos dos|So Paulo)\\b"
bn	<- "\\b(?i)(Brunei Darussalam|Brunei)\\b"
bg  <- "\\b(?i)(Bulgaria)\\b"
bf	<- "\\b(?i)(Burkina Faso)\\b"
bi	<- "\\b(?i)(Burundi)\\b"
kh	<- "\\b(?i)(Cambodia)\\b"
cm	<- "\\b(?i)(Cameroon)\\b"
ca  <- "\\b(?i)(Canada|Ontario(?!, CA)|Quebec|nova scotia|new brunswick(?![, N|,NJ| NJ])|manitoba|british columbia|, BC|prince edward island|saskatchewan|alberta|newfoundland|toronto|Vancouver(?![ WA| Washington| WA, USA|, WA, USA])|, ON|Winnipeg|, B.C.|,ON|(?<!Melbourne, )Victoria(?![ Australia|, Australia])|Burnaby|Fredericton|Montreal|Qubec)\\b"
cv	<- "\\b(?i)(Cape Verde)\\b"
ky	<- "\\b(?i)(Cayman Islands)\\b"
cf	<- "\\b(?i)(Central African Republic)\\b"
td	<- "\\b(?i)(Chad)\\b"
cl	<- "\\b(?i)(Chile)\\b"
cn  <- "\\b(?i)((?<!Hong Kong, )China|Beijing|GuangZhou|Shanghai|Tianjin|Shenzhen|Chengdu|Dongguan|Chongqing|Guangdong|ZhuHaiChina|Zhu haiChina|Zhengzhou|ZheJiang|HangZhou|zh-cn|(?<![Zurich|Lausanne]), CH|, CHN|Xiamen|Wuhan|beijingchina|Changsha|Hunan|ChengduChina|chinabeijing|ChinaShenzhen|^guang|^Hangz|Harbin Institute|Hefei|Hunan|^jiang|^Nanjing|^ShenZhen|GuangzhouChina|GuangZhouChina|NanjingChina|ShenzhenChina|XiamenChina|shenzhenChina|BeijingHaidian|wuhanChina)\\b"
cx	<- "\\b(?i)(Christmas Island)\\b"
cc	<- "\\b(?i)(Cocos Islands)\\b"
co	<- "\\b(?i)(Colombia|Bogataa|Medellin)\\b"
km	<- "\\b(?i)(Comoros)\\b"
cg	<- "\\b(?i)((?<!Democratic Republic of the )Congo)\\b"
cd	<- "\\b(?i)(Democratic Republic of the Congo|Brazzaville|Brazzaville|Brazzaville|Brazzaville, Congo, republic of|Congo - Brazzaville)\\b"
ck	<- "\\b(?i)(Cook Islands)\\b"
cr	<- "\\b(?i)(Costa Rica)\\b"
ci	<- "\\b(?i)(Côte dIvoire|Cote dIvoire|Cote d'Ivoire)\\b"
hr	<- "\\b(?i)(Croatia|Zagreb)\\b"
cu	<- "\\b(?i)(Cuba)\\b"
cw	<- "\\b(?i)(Curaçao)\\b"
cy	<- "\\b(?i)(Cyprus)\\b"
cz	<- "\\b(?i)(Czech Republic|Czech|Czechia|Prague|Brno|CzechRepublic)\\b"
dk	<- "\\b(?i)(Denmark|Copenhagen)\\b"
dj	<- "\\b(?i)(Djibouti)\\b"
dm	<- "\\b(?i)(Dominica)\\b"
do	<- "\\b(?i)(Dominican Republic|, DR|dominican|Dominican)\\b"
ec	<- "\\b(?i)(Ecuador)\\b"
eg	<- "\\b(?i)(Egypt)\\b"
sv	<- "\\b(?i)(El Salvador)\\b"
gq	<- "\\b(?i)(Equatorial Guinea)\\b"
er	<- "\\b(?i)(Eritrea)\\b"
ee	<- "\\b(?i)(Estonia)\\b"
et	<- "\\b(?i)(Ethiopia)\\b"
fk	<- "\\b(?i)(Falkland Islands|Malvinas)\\b"
fo	<- "\\b(?i)(Faroe Islands|Faroe)\\b"
fj	<- "\\b(?i)(Fiji)\\b"
fi	<- "\\b(?i)(Finland|Helsinki)\\b"
fr	<- "\\b(?i)(France|Paris|, FR|Toulouse|French Riviera|French Alps|Lille|Lyon|Sophia Antipolis)\\b"
gf	<- "\\b(?i)(French Guiana)\\b"
pf	<- "\\b(?i)(French Polynesia)\\b"
tf	<- "\\b(?i)(French Southern Territories)\\b"
ga	<- "\\b(?i)(Gabon)\\b"
gm	<- "\\b(?i)(Gambia)\\b"
# we decided to code instances of "georgia" as us since there was no easy way to demarcate and most seemed like they from the us 
ge	<- "\\b(?i)(Tbilisi|tbiilisi|Batumi|Kutaisi)\\b" #"\\b(?i)((?<!, )Georgia(?!, U)|Tbilisi)\\b"
de <- "\\b(?i)(Germany|German|Deutschland|Berlin|Bavaria|Bayern|Munich|Dresden|Frankfurt|Hamburg)\\b"
gh	<- "\\b(?i)(Ghana)\\b"
gi	<- "\\b(?i)(Gibraltar)\\b"
gr	<- "\\b(?i)(Greece)\\b"
gl	<- "\\b(?i)(Greenland)\\b"
gd	<- "\\b(?i)(Grenada)\\b"
gp	<- "\\b(?i)(Guadeloupe)\\b"
gu	<- "\\b(?i)(Guam)\\b"
gt	<- "\\b(?i)(Guatemala)\\b"
gg	<- "\\b(?i)(Guernsey)\\b"
gn	<- "\\b(?i)(Guinea)\\b"
gw	<- "\\b(?i)(Guinea-Bissau)\\b"
gy	<- "\\b(?i)(Guyana)\\b"
ht	<- "\\b(?i)(Haiti)\\b"
hm	<- "\\b(?i)(Heard Island|McDonald Islands)\\b"
va	<- "\\b(?i)(Holy See|Vatican)\\b"
hn	<- "\\b(?i)(Honduras)\\b"
hk	<- "\\b(?i)(Hong Kong)\\b"
hu	<- "\\b(?i)(Hungary|Budapest)\\b"
is	<- "\\b(?i)(Iceland)\\b"
id	<- "\\b(?i)(Indonesia|Yogyakarta|Bandung|Jakarta)\\b"
ir	<- "\\b(?i)(Iran)\\b"
iq	<- "\\b(?i)(Iraq)\\b"
ie	<- "\\b(?i)(Ireland|Dublin(?!, [OC]))\\b"
im	<- "\\b(?i)(Isle of Man)\\b"
il	<- "\\b(?i)(Israel)\\b"
it	<- "\\b(?i)(Italy|Rome(?!, GA)|Milan|Milano|Turin|Trieste|Bologna)\\b"
jm	<- "\\b(?i)(Jamaica(?! Plain))\\b"
jp  <- "\\b(?i)(Japan|Tokyo|Kyoto|Osaka|Yokohama|ja|ja_jp|ja-jp|Tsukuba)\\b"
jo	<- "\\b(?i)(Jordan(?!, U))\\b"
kz	<- "\\b(?i)(Kazakhstan)\\b"
ke	<- "\\b(?i)(Kenya|Nairobi)\\b"
ki	<- "\\b(?i)(Kiribati)\\b"
kp	<- "\\b(?i)(North Korea|Pyongyang)\\b"
kr	<- "\\b(?i)(South Korea|Seoul|Seuol|Daejeon|pangyo|Seongnam-si, Gyeonggi-do, Korea|Rep. of Korea|republic of korea|Korean)\\b"
xk  <- "\\b(?i)(Kosovo|Kosova|^Prishtina|^Prishtine|^Pristina)\\b"
kw	<- "\\b(?i)(Kuwait)\\b"
kg	<- "\\b(?i)(Kyrgyzstan)\\b"
la	<- "\\b(?i)(Lao Peoples Democratic Republic|Laos)\\b"
lv <- "\\b(?i)(Latvia)\\b"
lb	<- "\\b(?i)(Lebanon(?!, NH))\\b"
ls	<- "\\b(?i)(Lesotho)\\b"
lr	<- "\\b(?i)(Liberia)\\b"
ly	<- "\\b(?i)(Libya)\\b"
li	<- "\\b(?i)(Liechtenstein)\\b"
lu <- "\\b(?i)(Luxembourg)\\b"
lt <- "\\b(?i)(Lithuania)\\b"
mo	<- "\\b(?i)(Macao)\\b"
mk	<- "\\b(?i)(Macedonia)\\b"
mg	<- "\\b(?i)(Madagascar)\\b"
mw	<- "\\b(?i)(Malawi)\\b"
my	<- "\\b(?i)(Malaysia)\\b"
mv	<- "\\b(?i)(Maldives)\\b"
ml	<- "\\b(?i)(Mali)\\b"
mt <- "\\b(?i)(Malta)\\b"
mh	<- "\\b(?i)(Marshall Islands)\\b"
mq	<- "\\b(?i)(Martinique)\\b"
mr	<- "\\b(?i)(Mauritania)\\b"
mu	<- "\\b(?i)(Mauritius)\\b"
yt	<- "\\b(?i)(Mayotte)\\b"
mx	<- "\\b(?i)((?<!New )Mexico|(?<!New )Mxico|Zacatecas|Mex.|, MX|Villahermosa|Colima)\\b"
fm	<- "\\b(?i)(Micronesia)\\b"
md <- "\\b(?i)(Moldova)\\b"
mc <- "\\b(?i)(Monaco)\\b"
mn	<- "\\b(?i)(Mongolia)\\b"
me	<- "\\b(?i)(Montenegro)\\b"
ms	<- "\\b(?i)(Montserrat)\\b"
ma	<- "\\b(?i)(Morocco)\\b"
mz	<- "\\b(?i)(Mozambique)\\b"
mm	<- "\\b(?i)(Myanmar)\\b"
na	<- "\\b(?i)(Namibia)\\b"
nr	<- "\\b(?i)(Nauru)\\b"
np	<- "\\b(?i)(Nepal)\\b"
nl <- "\\b(?i)(Netherlands|(?<!rue d')Amsterdam|Holland(?!, [MO])|Zwolle|Willemstad|Curacao|Curaao|Utrecht|Breda|Eindhoven|Groningen)\\b"
nc	<- "\\b(?i)(New Caledonia)\\b"
nz	<- "\\b(?i)(New Zealand|Auckland|NZ|NZL)\\b"
ni	<- "\\b(?i)(Nicaragua)\\b"
ne	<- "\\b(?i)(Niger)\\b"
ng	<- "\\b(?i)(Nigeria|Lagos)\\b"
nu	<- "\\b(?i)(Niue)\\b"
nf	<- "\\b(?i)(Norfolk Island)\\b"
mp	<- "\\b(?i)(Northern Mariana Islands)\\b"
no	<- "\\b(?i)(Norway)\\b"
om	<- "\\b(?i)(Oman)\\b"
pk	<- "\\b(?i)(Pakistan)\\b"
pw	<- "\\b(?i)(Palau)\\b"
ps	<- "\\b(?i)(Palestine|Gaza)\\b"
pa	<- "\\b(?i)(Panama(?! City))\\b"
pg	<- "\\b(?i)(Papua New Guinea)\\b"
py	<- "\\b(?i)(Paraguay)\\b"
pe	<- "\\b(?i)(Peru|Arequipa)\\b"
ph	<- "\\b(?i)(Philippines)\\b"
pn	<- "\\b(?i)(Pitcairn)\\b"
pl <- "\\b(?i)(Poland|Warsaw|Wrocaw)\\b"
pt <- "\\b(?i)(Portugal|Lisboa|Lisbon|Azores|Porto(?! [Alegre|Velho])|Faro)\\b"
pr	<- "\\b(?i)(Puerto Rico)\\b"
qa	<- "\\b(?i)(Qatar)\\b"
re	<- "\\b(?i)(Réunion)\\b"
ro <- "\\b(?i)(Romania|Bucharest)\\b"
ru <- "\\b(?i)(Russia|Russian|Moscow(?!, I)|Yekaterinburg|St Petersburg(?![, F| FL| florida])|St. Petersburg(?![, F| FL| florida])|St-Petersburg(?!, F)|RU, |Simferopol|Sevastopol)\\b"
rw	<- "\\b(?i)(Rwanda)\\b"
bl	<- "\\b(?i)(Saint Barthélemy)\\b"
sh	<- "\\b(?i)(Saint Helena)\\b"
kn	<- "\\b(?i)(Saint Kitts|Nevis|Basseterre, St. Kitts)\\b"
lc	<- "\\b(?i)(Saint Lucia)\\b"
mf	<- "\\b(?i)(Saint Martin)\\b"
pm	<- "\\b(?i)(Saint Pierre|Miquelon)\\b"
vc	<- "\\b(?i)(Saint Vincent|Grenadines)\\b"
ws	<- "\\b(?i)(Samoa)\\b"
sm	<- "\\b(?i)(San Marino)\\b"
st	<- "\\b(?i)(Sao Tome|Principe)\\b"
sa	<- "\\b(?i)(Saudi Arabia)\\b"
sn <- "\\b(?i)(Senegal)\\b"
rs	<- "\\b(?i)(Serbia)\\b"
sc	<- "\\b(?i)(Seychelles)\\b"
sl	<- "\\b(?i)(Sierra Leone)\\b"
sg	<- "\\b(?i)(Singapore)\\b"
sx	<- "\\b(?i)(Sint Maarten)\\b"
sk <- "\\b(?i)(Slovak Republic|Slovakia)\\b"
si <- "\\b(?i)(Slovenia)\\b"
sb	<- "\\b(?i)(Solomon Islands)\\b"
so	<- "\\b(?i)(Somalia)\\b"
za	<- "\\b(?i)(South Africa|Cape Town|Johannesburg)\\b"
gs	<- "\\b(?i)(South Georgia|South Sandwich Islands|South Sandwich|Sandwich Islands)\\b"
ss	<- "\\b(?i)(South Sudan)\\b"
es <- "\\b(?i)(Spain|Madrid|Barcelona, Spain|Granada|Sevilla|Mallorca|Ibiza|Bilbao|Espana|Espaa|Zaragoza|Catalunya|Basque|Valencia(?!, [C|Ven])|Catalonia|Galicia)\\b"
lk	<- "\\b(?i)(Sri Lanka|Srilanka)\\b"
sd	<- "\\b(?i)(Sudan)\\b"
sr	<- "\\b(?i)(Suriname)\\b"
sj	<- "\\b(?i)(Svalbard|Jan Mayen)\\b"
sz	<- "\\b(?i)(Swaziland)\\b"
se <- "\\b(?i)(Sweden|Stockholm)\\b"
ch	<- "\\b(?i)(Switzerland|Zurich|Zrich|Lausanne)\\b"
sy	<- "\\b(?i)(Syrian Arab Republic|syria)\\b"
tw	<- "\\b(?i)(Taiwan|^Taipei)\\b"
tj	<- "\\b(?i)(Tajikistan)\\b"
tz	<- "\\b(?i)(Tanzania)\\b"
th	<- "\\b(?i)(Thailand)\\b"
tl	<- "\\b(?i)(Timor-Leste)\\b"
tg	<- "\\b(?i)(Togo)\\b"
tk	<- "\\b(?i)(Tokelau)\\b"
to	<- "\\b(?i)(Tonga)\\b"
tt	<- "\\b(?i)(Trinidad|Tobago)\\b"
tn	<- "\\b(?i)(Tunisia)\\b"
tr	<- "\\b(?i)(Turkey|Istanbul|Trkiye|Estambul)\\b"
tm	<- "\\b(?i)(Turkmenistan)\\b"
tc	<- "\\b(?i)(Turks|Caicos Islands)\\b"
tv	<- "\\b(?i)(Tuvalu)\\b"
ug	<- "\\b(?i)(Uganda)\\b"
ua	<- "\\b(?i)(Ukraine|Ukrane|Kyiv|Kiev|Zaporozhye|Dnepr|Dnipro)\\b"
ae	<- "\\b(?i)(United Arab|Emirates|UAE|Abu Dhabi)\\b"
gb <- "\\b(?i)(UK|UK,|United Kingdom|united-kingdom|London(?![, Ontario|, ON|Ontario| Canada|, Canada|, CA])|(?<!New )England|Wales|U.K.|Aberdeen|Cardiff(?!, CA)|Scotland|Edinburgh|Glasglow|Brighton(?!, [M|C]))\\b"
us	<- "\\b(?i)(?:Ala(?:(?:bam|sk)a)|American Samoa|Arizona|Arkansas|(?<!Baja )California|Colorado|Connecticut|Delaware|District of Columbia|Florida|(?<!Tbilisi, )Georgia(?!, tbilisi)|Guam|Hawaii|Idaho|Illinois|Indiana|Iowa|Kansas|Kentucky|Louisiana|Maine|Maryland|Massachusetts|Michigan|Minnesota|Miss(?:(?:issipp|our)i)|Montana|Nebraska|Nevada|New (?:Hampshire|Jersey|Mexico|York)|North (?:(?:Carolin|Dakot)a)|Ohio|Oklahoma|Oregon|Pennsylvania|Rhode Island|South (?:(?:Carolin|Dakot)a)|Tennessee|Texas|Utah|Vermont|Virgin(?:ia| Island(s?))|Washington|West Virginia|Wisconsin|Wyoming|United States|A[KLRSZ]|(?<![Ontario |Ontario, |Toronto |Toronto, |ON, |ON |Vancouver |Vancouver, |Vancouver,])C[AOT]|(?<![Berlin, |Hamburg, |Munich, |Dresden, |Frankfurt, ])DE|FL|G[AU]|HI|(?<![Jakarta, |Bandung, |Yogyakarta, |Pune, ])I[ALDN]|K[SY]|, LA|M[DEINOST]|(?<!Imperatriz-)MA|N[CDEHJMVY]|O[HK]|, OR|PA|RI|(?<!Blumenau, )S[CD](?![ - Brasil| - Brazil|, Brazil|, Brasil| Brazil| Brasil| - BR])|T[NX](?!, India)|UT|V[AIT]|W[AIVY](?!, Australia)|NYC|DC(?!, Colombia)|D.C.|US|USA|Seattle|San Francisco|Chicago|Philly|Philadelphia|Los Angeles|Portland|Houston|Phoenix|San Antonio|San Diego|Oakland|Boston|Bay Area|Berkeley|Stanford|Harvard|Dallas|Silicon|San Jose(?!, Costa Rica)|USC|Twin Cities|NYU|Yale|, Mass| Mass|newyork|new_york|U.S.|Atlanta|Austin|Albany|Manhattan|Brooklyn|Bronx|Cal|Cape Cod|Cleveland|Denver|Detroit|East Bay|Pittsburgh|Salt Lake|St. Louis|Honolulu|Miami|Minneapolis|Palo Alto|^san fran|socal|^SF|Merica$|Greater Nashville Area)\\b"
uy	<- "\\b(?i)(Uruguay)\\b"
uz	<- "\\b(?i)(Uzbekistan)\\b"
vu	<- "\\b(?i)(Vanuatu)\\b"
ve	<- "\\b(?i)(Venezuela)\\b"
vn	<- "\\b(?i)(Viet Nam|Vietnam|Vit Nam)\\b"
vg	<- "\\b(?i)(Virgin Islands, British|British Virgin Islands)\\b"
vi	<- "\\b(?i)(Virgin Islands, U.S.|Virgin Islands, US| US Virgin Islands)\\b"
wf	<- "\\b(?i)(Wallis and Futuna)\\b"
eh	<- "\\b(?i)(Western Sahara)\\b"
ye	<- "\\b(?i)(Yemen)\\b"
zm	<- "\\b(?i)(Zambia)\\b"
zw	<- "\\b(?i)(Zimbabwe)\\b"
asia <- "\\b(?i)(Asia)\\b"
africa <- "\\b(?i)((?<!South )Africa|Northern Africa|North Africa|West Africa|East Africa)\\b"
europe <- "\\b(?i)(^Europe|Central Europe|Eastern Europe|EU|Europa|European|Scandinavia)\\b"
americas <- "\\b(?i)(South America|Latin America|North America)\\b"
middle_east <- "\\b(?i)(EMEA|Middle East|eurasia)\\b"
planet_earth <- "\\b(?i)(Planet Earth|Worldwide|Earth planet|Earth, for now|Earth, Milky Way|Somewhere on Earth|Somewhere in the world|Earth|Earth, Solar System|Earth, Solar System, Milky Way|Anywhere, Earth|Somewhere in the world...|Somewhere on Planet Earth|Somewhere on the earth)\\b"
somewhere_in_space <- "\\b(?i)(deep space|Universe|Milky Way|Virgo Supercluster|ZZ9 Plural Z Alpha|Sector ZZ9 Plural Z Alpha|The Universe|milky way|Outer Space|3rd rock from the sun|The Milky Way|Galactic Sector ZZ9 Plural Z Alpha|Earth, Solar System, Milky Way, Laniakea|Laniakea Supercluster|Somewhere in the universe|Milkey Way|Middle-Earth|Milkyway Galaxy)\\b"
#some_any_nowhere <- "Somewhere|Nowhere|Anywhere|everywhere|Somewhere.|The middle of nowhere"
ficticious_places <- "\\b(?i)(Gensokyo|Matrix|The Matrix|Wonderland|Somewhere over the Rainbow|Somewhere under the rainbow|5th dimension|4th dimension)\\b"
worldwide_web <- "\\b(?i)(worldwide|international|WWW|cyberspace|Internets|Interwebz|interwebs)\\b"


# let's recode all of the strings and put them in a new data.table 
users_gh_cc <- data.table(oss_data) %>%
  # we want to recode these weird ones first so if a more specific code is available it will recode the entry as the most relevant 
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = asia), yes = "asia", no = "")) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = africa), 
                                      paste("africa", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = europe), 
                                      paste("europe", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = americas), 
                                      paste("americas", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = middle_east), 
                                      paste("middle_east", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = planet_earth), 
                                      paste("planet_earth", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = somewhere_in_space), 
                                      paste("somewhere_in_space", detected_country, sep=","), no = detected_country)) %>%
  #dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = some_any_nowhere), 
  #                                    paste("some_any_nowhere", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ficticious_places), 
                                      paste("ficticious_places", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = worldwide_web), 
                                      paste("worldwide_web", detected_country, sep=","), no = detected_country)) %>%
  # now we are adding all of the country codes 
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = af), paste("af", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ax), paste("ax", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = al), paste("al", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = dz), paste("dz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = as), paste("as", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ad), paste("ad", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ao), paste("ao", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ai), paste("ai", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = aq), paste("aq", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ag), paste("ag", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ar), paste("ar", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = am), paste("am", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = aw), paste("aw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = au), paste("au", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = at), paste("at", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = be), paste("be", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bg), paste("bg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = br), paste("br", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = az), paste("az", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bh), paste("bh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bs), paste("bs", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bd), paste("bd", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bb), paste("bb", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = by), paste("by", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bz), paste("bz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bj), paste("bj", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bm), paste("bm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bt), paste("bt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bo), paste("bo", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bq), paste("bq", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ba), paste("ba", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bw), paste("bw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bv), paste("bv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bn), paste("bn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bf), paste("bf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bi), paste("bi", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cy), paste("cy", detected_country, sep=","), no = detected_country)) %>% 
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cz), paste("cz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cn), paste("cn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kh), paste("kh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cm), paste("cm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ca), paste("ca", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cv), paste("cv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ky), paste("ky", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cf), paste("cf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = td), paste("td", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cl), paste("cl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cx), paste("cx", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cc), paste("cc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = co), paste("co", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = km), paste("km", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ck), paste("ck", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cr), paste("cr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ci), paste("ci", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = hr), paste("hr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cu), paste("cu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = cw), paste("cw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = dk), paste("dk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ee), paste("ee", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fi), paste("fi", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fr), paste("fr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = dj), paste("dj", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = dm), paste("dm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = do), paste("do", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ec), paste("ec", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = eg), paste("eg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sv), paste("sv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gq), paste("gq", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = er), paste("er", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = et), paste("et", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fk), paste("fk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fo), paste("fo", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fj), paste("fj", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = de), paste("de", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gr), paste("gr", detected_country, sep=","), no = detected_country)) %>% 
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gf), paste("gf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pf), paste("pf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tf), paste("tf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ga), paste("ga", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gm), paste("gm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ge), paste("ge", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gh), paste("gh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gi), paste("gi", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gl), paste("gl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gd), paste("gd", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gp), paste("gp", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gu), paste("gu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gt), paste("gt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gg), paste("gg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gn), paste("gn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gw), paste("gw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gy), paste("gy", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = hu), paste("hu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ie), paste("ie", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = it), paste("it", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ht), paste("ht", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = hm), paste("hm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = va), paste("va", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = hn), paste("hn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = hk), paste("hk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = is), paste("is", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = id), paste("id", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ir), paste("ir", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = iq), paste("iq", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = im), paste("im", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = il), paste("il", detected_country, sep=","), no = detected_country)) %>%
  # i did india down here so that i didn't have to name a variable "in", which created problems 
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = "\\b(?i)(India|Mumbai|Delhi|Kolkata|Chennai|Bangalore|Hyderabad|Ahmedabad|Pune|VIT (?!Nam)|NIT |Amritsar|(?<!Lahore[ ,])Punjab(?! Pakistan)|Bhubaneswar|Gandhinagar|Gujarat|^IIT |Indian|Jaipur|jammu|kashmir|^MNNIT)\\b"), paste("in", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = jp), paste("jp", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lv), paste("lv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lu), paste("lu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lt), paste("lt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = jm), paste("jm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = jo), paste("jo", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kz), paste("kz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ke), paste("ke", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ki), paste("ki", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kp), paste("kp", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kr), paste("kr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kw), paste("kw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kg), paste("kg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = xk), paste("xk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = la), paste("la", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lb), paste("lb", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ls), paste("ls", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lr), paste("lr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ly), paste("ly", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = li), paste("li", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mt), paste("mt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mo), paste("mo", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mk), paste("mk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mg), paste("mg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mw), paste("mw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = my), paste("my", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mv), paste("mv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ml), paste("ml", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mh), paste("mh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mq), paste("mq", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mr), paste("mr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mu), paste("mu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mx), paste("mx", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = yt), paste("yt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = fm), paste("fm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = md), paste("md", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mc), paste("mc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mn), paste("mn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = me), paste("me", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ms), paste("ms", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ma), paste("ma", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mz), paste("mz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mm), paste("mm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = na), paste("na", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nr), paste("nr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = np), paste("np", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nl), paste("nl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nc), paste("nc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nz), paste("nz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ni), paste("ni", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ne), paste("ne", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ng), paste("ng", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nu), paste("nu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = nf), paste("nf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mp), paste("mp", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = no), paste("no", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = om), paste("om", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pk), paste("pk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pw), paste("pw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ps), paste("ps", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pa), paste("pa", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pg), paste("pg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = py), paste("py", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pe), paste("pe", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ph), paste("ph", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pn), paste("pn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pl), paste("pl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pt), paste("pt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pr), paste("pr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = qa), paste("qa", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = re), paste("re", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ro), paste("ro", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ru), paste("ru", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = rw), paste("rw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = bl), paste("bl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sh), paste("sh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = kn), paste("kn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lc), paste("lc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = mf), paste("mf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = pm), paste("pm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = vc), paste("vc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ws), paste("ws", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sm), paste("sm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = st), paste("st", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sa), paste("sa", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sn), paste("sn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = rs), paste("rs", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sc), paste("sc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sl), paste("sl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sg), paste("sg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sx), paste("sx", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sk), paste("sk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = si), paste("si", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sb), paste("sb", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = so), paste("so", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = za), paste("za", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gs), paste("gs", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ss), paste("ss", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = es), paste("es", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = lk), paste("lk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sd), paste("sd", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sr), paste("sr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sj), paste("sj", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sz), paste("sz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = se), paste("se", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ch), paste("ch", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = sy), paste("sy", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tw), paste("tw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tj), paste("tj", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tz), paste("tz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = th), paste("th", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tl), paste("tl", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tg), paste("tg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tk), paste("tk", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = to), paste("to", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tt), paste("tt", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tn), paste("tn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tr), paste("tr", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tm), paste("tm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tc), paste("tc", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = tv), paste("tv", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ug), paste("ug", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ua), paste("ua", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ae), paste("ae", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = gb), paste("gb", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = uy), paste("uy", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = uz), paste("uz", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = vu), paste("vu", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ve), paste("ve", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = vn), paste("vn", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = vg), paste("vg", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = vi), paste("vi", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = wf), paste("wf", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = eh), paste("eh", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = ye), paste("ye", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = zm), paste("zm", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = zw), paste("zw", detected_country, sep=","), no = detected_country)) %>%
  dt_mutate(detected_country = ifelse(test = str_detect(string = oss_data$location, pattern = us), paste("us", detected_country, sep=","), no = detected_country)) %>% 
  dplyr::rename(new_country_code = detected_country)

# removing the commas at the end of the strings from the column just created
users_gh_cc$new_country_code <- gsub("\\,$", "", users_gh_cc$new_country_code)
users_gh_cc$new_country_code <- trimws(users_gh_cc$new_country_code)

users_gh_cc %>% 
  select(login, new_country_code, everything()) %>% 
  drop_na(new_country_code)
```

Ok, now we need to see if we can increase the number of users classified based on the other column data. 

First, let's try imputing old country_code information into new_county_code information, cleaning spaces that were misclassified, and renaming common multiple country mistakes. This cleaning step actually lowers the count because of false-positive spaces that were counted as a country code.  

```{r}
# after recoding all the strings into new_cc, lets impute old_cc into new_cc when missing 
users_with_old_cc <- as_tibble(users_gh_cc) %>% 
  select(login, new_country_code, country_code, everything()) %>% 
  mutate(new_country_code = ifelse(is.na(new_country_code), country_code, new_country_code),
         new_country_code = ifelse(new_country_code == "", country_code, new_country_code),
         country_code = trimws(country_code),
         new_country_code = trimws(new_country_code),
            # fixing systematic false positives (would be better if i did this in regex but it would take a long time)
         new_country_code = ifelse(test = str_detect(string = new_country_code, pattern = "us,br"), yes = "br", no = new_country_code),
         new_country_code = ifelse(test = str_detect(string = new_country_code, pattern = "us,in"), yes = "in", no = new_country_code))

users_with_old_cc %>% 
  drop_na(new_country_code)

multiple_cc_fixes <- users_with_old_cc %>% 
  dt_filter(grepl(",", new_country_code)) %>% 
  #separate_rows(new_country_code, sep = "," , convert = FALSE) %>% 
  mutate(check = str_replace_all(new_country_code, "planet_earth", ""),
         check = str_replace_all(check, "worldwide_web", ""),
         check = str_replace_all(check, "somewhere_in_space", ""),
         check = str_replace_all(check, "europe", ""),
         check = str_replace_all(check, "asia", ""),
         check = str_replace_all(check, "americas", ""),
         check = str_replace_all(check, "africa", ""),
         check = str_replace_all(check, "middle_east", ""),
         check = str_replace_all(check, "ficticious_places", ""),
         check = str_replace_all(check, "\\,$", ""),
         check = str_replace_all(check, "\\,$", ""),
         check = str_replace_all(check, "\\,$", "")) %>% 
  #mutate(check = ifelse(str_detect(new_country_code, "\\b(\\,)\\b"), "us", "bleh")) %>% 
  select(login, check) 

users_with_old_cc <- users_with_old_cc %>% 
  left_join(multiple_cc_fixes, by = "login") %>% 
  mutate(new_country_code = ifelse(is.na(check), new_country_code, check)) %>% 
  mutate(new_country_code = na_if(new_country_code, "")) %>% select(-check) 

users_with_old_cc %>% 
  drop_na(new_country_code)

# around 5,000 cases that do not match 
non_matching <- users_with_old_cc %>% 
  dt_mutate(match_test = ifelse(country_code == new_country_code, TRUE, FALSE)) %>% 
  dt_filter(match_test == FALSE) %>% 
  select(login, new_country_code, country_code, location, city, match_test, everything())
non_matching # there is some room to improve with these (might just impute original country code instead)

non_matching %>% 
  count(new_country_code) %>% 
  arrange(-n)

# dropped from 2,078 that do not match (bc we allow multiple countries) to ~540 after recoding 
non_matching %>% 
  select(login, country_code, new_country_code, everything()) %>% 
  dt_filter(grepl(",", new_country_code)) #%>% count(new_country_code) %>% arrange(-n)

```

Now, let's see if the sectored academic data (with country codes) help classify more users.

```{r}
# joining increases dataset from 717492 to 717558
users_with_uni_cc <- users_with_old_cc %>% 
  left_join(academic_data, by = "login") %>% 
  #select(login, country, country_code, new_country_code, institution) %>% 
  #filter(is.na(new_country_code) & !is.na(country)) %>% 
  mutate(tmp_cc = countrycode(country, origin = 'country.name', destination = 'iso2c'),
         tmp_cc = tolower(tmp_cc),
         new_country_code = ifelse(is.na(new_country_code), tmp_cc, new_country_code)) %>% 
  select(login, new_country_code, country_code, everything(), 
         -institution, -tmp_cc, is_academic, -country, -is_academic)

users_with_uni_cc %>% 
  count(new_country_code) %>% 
  arrange(-n)

users_with_uni_cc %>% 
  drop_na(new_country_code)
# we are up from 700,105 users in the first draft of the Scientific Reports paper to 733,399 users 
```

Next, let's try emails. This entails pulling out all of the country codes from the email subdomains and then classifying into countries from there.

```{r}
email_domain_by_country <- read_csv("/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/government/email_to_country_code.csv")
email_domain_by_country <- email_domain_by_country %>% 
  mutate(country_code = tolower(country_code)) %>% 
  rename(domain = country_email, 
         cc_by_email = country_code) %>% 
  select(domain, cc_by_email, country_name); email_domain_by_country

cc_by_email <- users_with_uni_cc %>% 
  filter(is.na(new_country_code) & !is.na(email)) %>% 
  dt_mutate(domain = str_extract(email, ".[^.]*$")) %>% 
  filter(domain != ".com" & domain != ".org" & domain != ".io" & domain != ".im" &
         domain != ".net" & domain != ".me" & domain != ".cc" & domain != ".co" & domain != ".sh") %>% 
  left_join(email_domain_by_country, by = "domain") %>% 
  drop_na(cc_by_email) %>% 
  select(login, cc_by_email) 

users_with_email_cc <- users_with_uni_cc %>% 
  left_join(cc_by_email, by = "login") %>% 
  mutate(new_country_code = ifelse(is.na(new_country_code), cc_by_email, new_country_code))

users_with_email_cc <- users_with_email_cc %>% 
  mutate(domain = str_extract(email, ".[^.]*$"), 
         alt_domains = ifelse(str_detect(domain, ".gov"), "us", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".mil"), "us", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".nyc"), "us", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".army"), "us", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".eu"), "europe", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".asia"), "asia", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".quebec"), "ca", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".berlin"), "de", new_country_code),
         alt_domains = ifelse(str_detect(domain, ".tokyo"), "jp", new_country_code),
         jp_cn = ifelse(str_detect(company, "china"), "cn", new_country_code),
         jp_cn = ifelse(str_detect(company, "japan"), "jp", new_country_code),
         new_country_code = ifelse(is.na(new_country_code), alt_domains, new_country_code),
         new_country_code = ifelse(is.na(new_country_code), jp_cn, new_country_code)) %>% 
  select(login, new_country_code, everything(), -cc_by_email, -alt_domains, -jp_cn)

users_with_email_cc %>% 
  count(new_country_code) %>% 
  arrange(-n)

users_with_email_cc %>% 
  drop_na(new_country_code)
# we are up from 700,105 users in the first draft of the Scientific Reports paper to 751,222 users 
```

```{r}
users_with_email_cc %>% 
  filter(is.na(new_country_code) & !is.na(email)) %>% 
  dt_mutate(domain = str_extract(email, ".[^.]*$")) %>% 
  filter(domain != ".com" & domain != ".org" & domain != ".io" & domain != ".im" &
         domain != ".net" & domain != ".me" & domain != ".cc" & domain != ".co" & domain != ".sh") %>% 
  left_join(email_domain_by_country, by = "domain") %>% 
  filter(domain == ".edu") %>% 
  select( email, company, login, everything())
# also about 550 .edu people that should be classified in academia
# most are cases where the put info in the company column but still have .edu domains 
# too time consuming to pursue this here, but could develop in academia 

# could improve on this -- about 30-40 rows of random cities without country codes
users_with_email_cc %>% 
  filter(is.na(new_country_code) & !is.na(city)) %>% 
  count(location) %>% 
  arrange(-n)

# also 200 rows with lat and lon but no country code  
users_with_email_cc %>% 
  filter(is.na(new_country_code) & !is.na(long) & long != 0) %>% 
  count(location) %>% 
  arrange(-n)
```

First, we want to get the total counts of countries even when they are listed more than once.

```{r}
users_with_email_cc %>% 
  # separate multiple countries 
  separate_rows(new_country_code, sep = "," , convert = FALSE) %>% 
  # trim ws 
  dplyr::mutate(new_country_code = trimws(new_country_code)) %>% 
  # remove missing data 
  dplyr::filter(!is.na(new_country_code)) %>% 
  # rename to country_code 
  dplyr::mutate(country_code = new_country_code) %>% 
  # count and arrange by n 
  dplyr::count(country_code) %>% arrange(-n)
```

Here, we see that the US, China, the UK, Germany and India have the highest totals, which aligns with what we found in our original cleaning documents. 

Next, this next snippet removes all of the missing data, imputes the country codes from the gh_torrent data when our recode didn't capture the country code, and renames the column to "country_code" so that others have an easier time interpreting the data on the database. Note that we end up with a different row count then our last snippet because we did not separate the logins with multiple countries 

```{r}
users_with_countries <- users_with_email_cc

users_with_countries <- users_with_countries %>% 
  dplyr::mutate(new_country_code = trimws(new_country_code)) %>%
  drop_na(new_country_code) %>% 
  dplyr::select(login, new_country_code) %>%
  dplyr::rename(country_code = new_country_code) %>% 
  dplyr::mutate(country_code_vis = ifelse(test = str_detect(string = country_code,
                                      pattern = ","), paste("multiple"), no = country_code)) %>% 
  dplyr::mutate(country_code_di = ifelse(test = str_detect(string =  country_code,
                                      pattern = ","), paste("multiple"), no = "single")) %>% 
  dplyr::select(login, country_code, country_code_di, country_code_vis)

# trimming white space (just in case)
users_with_countries$country_code <- stringi::stri_trim_both(users_with_countries$country_code)
users_with_countries$country_code_di <- stringi::stri_trim_both(users_with_countries$country_code_di)
users_with_countries$country_code_vis <- stringi::stri_trim_both(users_with_countries$country_code_vis)

# country counts after re-coding (without separating)
users_with_countries %>%
  dplyr::filter(!is.na(country_code)) %>% 
  dplyr::count(country_code, sort = TRUE)
```
```{r}
# unique users 
length(users_with_countries$login)
```
```{r}
# number of users with multiple country codes 
users_with_countries %>% 
  dplyr::filter(grepl(",", country_code)) %>% 
  count()
```
```{r}
users_with_countries %>% dplyr::filter(grepl(",", country_code)) %>% count() / length(users_gh_cc$login)
```
```{r}
exclusion_codes <- c("europe", "asia", "africa", "americas", "middle_east", "planet_earth", 
                     "worldwide_web", "somewhere_in_space", "ficticious_places") 
users_with_countries %>% dplyr::filter(country_code %in% exclusion_codes) %>% count()
users_with_countries %>% dplyr::filter(country_code %in% exclusion_codes) %>% count() / length(users_gh_cc$login)
```
```{r}
all_excluded <- c("europe", "asia", "africa", "americas", "middle_east", "planet_earth", 
                     "worldwide_web", "somewhere_in_space", "ficticious_places")  
users_with_countries %>% dplyr::filter(country_code %in% all_excluded | grepl(",", country_code)) %>% count()
users_with_countries %>% dplyr::filter(country_code %in% all_excluded | grepl(",", country_code)) %>% count() / length(users_gh_cc$login)
```
```{r}
users_with_countries <- users_with_countries %>% 
  mutate(country_code_vis = na_if(country_code_vis, "europe"),
         country_code_vis = na_if(country_code_vis, "asia"),
         country_code_vis = na_if(country_code_vis, "africa"),
         country_code_vis = na_if(country_code_vis, "americas"),
         country_code_vis = na_if(country_code_vis, "middle_east"),
         country_code_vis = na_if(country_code_vis, "planet_earth"),
         country_code_vis = na_if(country_code_vis, "worldwide_web"),
         country_code_vis = na_if(country_code_vis, "somewhere_in_space"),
         country_code_vis = na_if(country_code_vis, "ficticious_places"))

users_with_countries <- users_with_countries %>% 
  mutate(country_name = countrycode(country_code_vis, origin = 'iso2c', destination = 'country.name'),
         country_name = ifelse(str_detect(country_code_vis, "xk"), "Kosovo", country_name),
         country_name = ifelse(str_detect(country_code_vis, "multiple"), "Multiple Countries", country_name))

users_with_countries
```

```{r}
intl_totals_multiples_separated <- users_with_countries %>% 
  separate_rows(country_code, sep = "," , convert = FALSE) %>% 
  dplyr::mutate(country_code = trimws(country_code)) %>%
  dplyr::count(country_code) %>% 
  arrange(-n)
intl_totals

write_csv(intl_totals_multiples_separated, "/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/intl_totals/intl_totals_multiples_separated.csv")

intl_totals_multiples_included <- users_with_countries %>%
  dplyr::filter(!is.na(country_code_vis)) %>% 
  dplyr::count(country_code_vis, sort = TRUE)
intl_totals_multiples_included

write_csv(intl_totals_multiples_included, "/sfs/qumulo/qhome/kb7hp/git/oss-2020/data/intl_totals/intl_totals_multiples_included.csv")
```

Now that the countries have been recoded, let's write that table to PostgreSQL. Note that this *could* be used as the nodelist for the networks, but we actually want to add in more node attributes a little later on. 

```{r writing filtered gh_users data back to database}
# reconnecting to the database 
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# writing the new users_gh_cc table to postgis_2
dbWriteTable(conn, c("gh", "sna_ctr_ctry_codes"), users_with_countries, row.names = FALSE)

# disconnect from postgresql database  
dbDisconnect(conn)
```
